Move BABE testsuite to its own file
